refactor(command-feature): rework the architecture
