+++ src/features/discord/messages/services/command/feature/discord-message-command-feature.service.ts
@@ -12,7 +12,7 @@ import { DiscordMessageConfigService } from "../../config/discord-message-config
 import { DiscordMessageCommandFeatureEmptyContentErrorService } from "./services/discord-message-command-feature-empty-content-error.service";
 import { DiscordMessageCommandFeatureEmptyFeatureNameErrorService } from "./services/discord-message-command-feature-empty-feature-name-error.service";
 import { DiscordMessageCommandFeatureWrongFeatureNameErrorService } from "./services/discord-message-command-feature-wrong-feature-name-error.service";
-import { DiscordMessageCommandFeatureNoonService } from "./services/noon/discord-message-command-feature-noon.service";
+import { DiscordMessageCommandFeatureNoonService } from "./features/noon/services/discord-message-command-feature-noon.service";
 
 export class DiscordMessageCommandFeatureService extends AbstractService {
   private static _instance: DiscordMessageCommandFeatureService;
