+++ src/features/discord/messages/services/command/feature/features/noon/services/discord-message-command-feature-noon.service.spec.ts
@@ -1,10 +1,10 @@
-import { ServiceNameEnum } from "../../../../../../../../enums/service-name.enum";
-import { CoreEventService } from "../../../../../../../core/services/core-event.service";
-import { IAnyDiscordMessage } from "../../../../../types/any-discord-message";
-import { DiscordMessageCommandFeatureNameEnum } from "../../enums/discord-message-command-feature-name.enum";
+import { ServiceNameEnum } from "../../../../../../../../../enums/service-name.enum";
+import { CoreEventService } from "../../../../../../../../core/services/core-event.service";
+import { IAnyDiscordMessage } from "../../../../../../types/any-discord-message";
+import { DiscordMessageCommandFeatureNameEnum } from "../../../enums/discord-message-command-feature-name.enum";
 import { DiscordMessageCommandFeatureNoonService } from "./discord-message-command-feature-noon.service";
 
-jest.mock(`../../../../../../../logger/services/chalk/chalk.service`);
+jest.mock(`../../../../../../../../logger/services/chalk/chalk.service`);
 
 describe(`DiscordMessageCommandFeatureNoonService`, (): void => {
   let service: DiscordMessageCommandFeatureNoonService;
