+++ src/features/discord/messages/services/command/feature/features/noon/services/discord-message-command-feature-noon.service.ts
@@ -1,10 +1,10 @@
 import _ from "lodash";
-import { AbstractService } from "../../../../../../../../classes/abstract.service";
-import { ServiceNameEnum } from "../../../../../../../../enums/service-name.enum";
-import { IDiscordMessageResponse } from "../../../../../interfaces/discord-message-response";
-import { IAnyDiscordMessage } from "../../../../../types/any-discord-message";
-import { DiscordMessageCommandFeatureNameEnum } from "../../enums/discord-message-command-feature-name.enum";
-import { IDiscordMessageCommandFeatureNameNoon } from "../../types/discord-message-command-feature-name-noon";
+import { AbstractService } from "../../../../../../../../../classes/abstract.service";
+import { ServiceNameEnum } from "../../../../../../../../../enums/service-name.enum";
+import { IDiscordMessageResponse } from "../../../../../../interfaces/discord-message-response";
+import { IAnyDiscordMessage } from "../../../../../../types/any-discord-message";
+import { DiscordMessageCommandFeatureNameEnum } from "../../../enums/discord-message-command-feature-name.enum";
+import { IDiscordMessageCommandFeatureNameNoon } from "../types/discord-message-command-feature-name-noon";
 
 export class DiscordMessageCommandFeatureNoonService extends AbstractService {
   private static _instance: DiscordMessageCommandFeatureNoonService;
