+++ src/features/discord/messages/services/command/feature/features/noon/types/discord-message-command-feature-name-noon.ts
@@ -1,4 +1,4 @@
-import { DiscordMessageCommandFeatureNameEnum } from "../enums/discord-message-command-feature-name.enum";
+import { DiscordMessageCommandFeatureNameEnum } from "../../../enums/discord-message-command-feature-name.enum";
 
 export type IDiscordMessageCommandFeatureNameNoon =
   | DiscordMessageCommandFeatureNameEnum.NOON
